scripts/requirements.txt
google-cloud-bigquery
requests
PyGithub
